+++ CHANGES.md
@@ -17,8 +17,10 @@
   including the protocols "s3" and "file". It also allows patching
   xcube multi-level datasets (`*.levels` format).
   
-* In configuration for `xcube server`, datasets defined in datastores may now
-  have user-set identifiers, in case their path unambiguously defines a dataset
+* In the configuration for `xcube server`, datasets defined in `DataStores` 
+  may now have user-defined identifiers. In case the path does not unambiguously 
+  define a dataset (because, e.g., it contains wildcards, providing a 
+  user-defined identifier will raise an error. 
 
 ### Fixes